requirements.txt
python-telegram-bot==20.7
apscheduler==3.10.4
pytz
nest_asyncio
deep-translator==1.11.4
